Split Demo.cpp into multiple files
Tons of stuff going on and file is getting too big, split off the responsibility into more classes/files.
