+++ Demo/Demo.vcxproj.filters
@@ -1,5 +1,20 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup />
-  <ItemGroup />
+  <ItemGroup>
+    <ClCompile Include="Demo.cpp" />
+    <ClCompile Include="Sound.cpp" />
+    <ClCompile Include="RenderBuilder.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Demo.h" />
+    <ClInclude Include="TypeToString.h" />
+    <ClInclude Include="Character.h" />
+    <ClInclude Include="MoveableCharacter.h" />
+    <ClInclude Include="AiCharacter.h" />
+    <ClInclude Include="Sound.h" />
+    <ClInclude Include="RenderBuilder.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <ResourceCompile Include="Demo.rc" />
+  </ItemGroup>
 </Project>
