+++ Demo/Sound.cpp
@@ -1,0 +1,8 @@
+#include "Sound.h"
+
+Bat::ISoundEngine* snd;
+
+void InitializeSound()
+{
+	snd = Bat::Audio::CreateSoundPlaybackDevice();
+}
