+++ Demo/Sound.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "Audio.h"
+
+extern Bat::ISoundEngine* snd;
+
+void InitializeSound();
