+++ Engine/Events/Event.h
@@ -1,5 +1,12 @@
 #pragma once
 
+#include <functional>
+#include <typeindex>
+#include <unordered_map>
+#include <array>
+#include <vector>
+#include "BatAssert.h"
+
 // TODO: Make this thread safe so we can listen for events on other threads
 namespace Bat
 {
